Fix query parsing and allow more types
